repo setup and uncompressed baseline
Needed to make a few changes to mobilenet-v2 for it to work on CIFAR-10.

Since MobileNet-v2 was designed for 224x224 images, it downsamples the
image fast downto a small patch. If we adopt this for CIFAR-10's 32x32
images, we will have our classifier layer wrestling with just 1-5
pixels.

To fix this, we change the stride in two places.

As for transforms, we use the standard normalization, followed by
cutout - which is basically like dropout, but for the source image
itself.

I got 94% accuracy with the random seed 6886.